Split `no4` `e_secrets` file
