+++ config/splat.pspeu.stno4.yaml
@@ -46,7 +46,8 @@ segments:
       - [0x4A0, c, no4_psp/unk_4A0]
       - [0xEA0, c, no4/st_debug]
       - [0x11E0, c, no4/e_killer_fish]
-      - [0x1798, c, no4_psp/unk_1798]
+      - [0x1798, c, no4_psp/e_secrets]
+      - [0x2108, c, no4_psp/unk_2108]
       - [0xC388, c, no4/create_entity]
       - [0xD330, c, no4/e_bone_archer]
       - [0xE4E8, c, no4_psp/e_stage_name]
@@ -96,7 +97,7 @@ segments:
       - [0x68270, .data, no4/e_skeleton_ape]
       - [0x68430, .data, no4/e_spear_guard]
       - [0x68710, data]
-      - [0x68A80, .rodata, no4_psp/unk_1798]
+      - [0x68A80, .rodata, no4_psp/unk_2108]
       - [0x68AF8, .rodata, no4/e_red_door]
       - [0x68B10, .rodata, no4/e_fishhead]
       - [0x68B28, .rodata, no4/e_toad]
