+++ src/st/no4_psp/e_secrets.c
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "common.h"
+
+INCLUDE_ASM("st/no4_psp/psp/no4_psp/e_secrets", func_pspeu_09238E18);
+
+INCLUDE_ASM("st/no4_psp/psp/no4_psp/e_secrets", func_pspeu_092391D8);
+
+INCLUDE_ASM("st/no4_psp/psp/no4_psp/e_secrets", func_pspeu_09239568);
