--- src/st/no4_psp/unk_1798.c
@@ -1,100 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "common.h"
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09238E18);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092391D8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09239568);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09239788);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092398B8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09239B00);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923B288);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923B378);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923BAB0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923BB30);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923C460);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D0B8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D0C8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D180);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D428);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D4A0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D568);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923D9E0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923DD38);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923E368);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923E690);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923EA80);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923EED0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923F390);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923F860);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923FB88);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_0923FF60);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09240350);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09240680);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092409A8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09240CD0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09241098);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09241410);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09241870);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09241B28);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09241E38);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09242200);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092423C8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092427C8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092428E0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09242940);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09242AD0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09242D08);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09242E30);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09242FB8);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09243260);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_092432C0);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09243320);
-
-INCLUDE_ASM("st/no4_psp/psp/no4_psp/unk_1798", func_pspeu_09243478);
